Implementing galaxy rotation
Former-commit-id: 0dc3cd50d737dc05ca5462b9cb3a73318879355d [formerly 196a595d2e5f5727923c36169d4d32db6402dea8]
Former-commit-id: 9d4b5d4e3dea85316deec93a2b7668c7763755c3

+++ Renderer/Renderer/Skybox.cpp
@@ -95,7 +95,7 @@ unsigned int Skybox::LoadCubemap(const std::vector<std::string> & faces)
     glBindTexture(GL_TEXTURE_CUBE_MAP, texture_id);
 
     int width, height, nrChannels;
-    unsigned int face_size = faces.size();
+    unsigned int face_size = static_cast<unsigned int>(faces.size());
     for (unsigned int i = 0; i < face_size; ++i)
     {
         unsigned char * data = stbi_load(faces[i].c_str(), &width, &height, &nrChannels, 0);
